Move Planning controller
